pyproject.toml
[tool.poetry]
name = "diplom"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.2.4"
python-dotenv = "^1.1.1"
psycopg2-binary = "^2.9.10"
pillow = "^11.3.0"
ipython = "^9.4.0"
requests = "^2.32.4"
djangorestframework = "^3.16.0"


[tool.poetry.group.dev.dependencies]
isort = "^6.0.1"


[tool.poetry.group.lint.dependencies]
black = "^25.1.0"
isort = "^6.0.1"
flake8 = "^7.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
